Update README.md
This is my very first web development project where I created a simple HTML page to promote a 55th birthday celebration. This project marks the beginning of my Full Stack Web Development journey, and I built it as part of the course from Angela Yu on Udemy.

Project Features
A heading displaying the celebration title and the date.
A celebratory image of the number 55 made from balloons.
A list of things to bring to the event.
The event's location, with a link to Google Maps for directions.
Technologies Used
HTML for structuring the webpage content.
Basic CSS styling.